backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-migrate
flask-jwt-extended
python-dotenv
passlib
bcrypt
